users cant see managers

+++ app/helpers/application_helper.rb
@@ -1,10 +1,12 @@
 module ApplicationHelper
 
 	def link_for_user user
-		if current_user.friends_with?(user)
-			button_to 'Remove friend', friendship_path(user), method: :delete
-		else
-			link_to 'Send request', send_request_friendship_path(user)
+		if !current_user.admin
+			if current_user.friends_with?(user)
+				button_to 'Remove friend', friendship_path(user), method: :delete
+			else
+				link_to 'Send request', send_request_friendship_path(user)
+			end
 		end
 	end
 
@@ -38,5 +40,5 @@ module ApplicationHelper
 			end
 		end
 	end
-	
+
 end
